Integrate Authentication with React-Adal library
