+++ package.json
@@ -7,6 +7,7 @@
     "enzyme-adapter-react-16": "^1.5.0",
     "prop-types": "^15.6.2",
     "react": "^16.4.2",
+    "react-adal": "^0.4.18",
     "react-dom": "^16.4.2",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.5"
